+++ railgun/cmd/rootcmd/admission.go
@@ -3,8 +3,11 @@ package rootcmd
 import (
 	"context"
 
+	"github.com/bombsimon/logrusr"
+	"github.com/go-logr/logr"
 	"github.com/kong/kubernetes-ingress-controller/pkg/admission"
 	"github.com/kong/kubernetes-ingress-controller/pkg/util"
+	"github.com/kong/kubernetes-ingress-controller/railgun/internal/diagnostics"
 	"github.com/kong/kubernetes-ingress-controller/railgun/pkg/config"
 )
 
@@ -40,5 +43,23 @@ func StartAdmissionServer(ctx context.Context, c *config.Config) error {
 		err := srv.ListenAndServeTLS("", "")
 		log.WithError(err).Error("admission webhook server stopped")
 	}()
+	return nil
+}
+
+func StartProfilingServer(ctx context.Context, c *config.Config) error {
+	deprecatedLogger, err := util.MakeLogger(c.LogLevel, c.LogFormat)
+
+	var logger logr.Logger = logrusr.NewLogger(deprecatedLogger)
+	if err != nil {
+		return err
+	}
+	if c.EnableProfiling {
+		s := diagnostics.Server{Logger: logger}
+		go func() {
+			if err := s.Listen(ctx); err != nil {
+				logger.Error(err, "unable to start diagnostics server")
+			}
+		}()
+	}
 	return nil
 }